+++ _ide_test_container.php
@@ -7,13 +7,31 @@ use Raptor\TestUtils\TestDataContainer\TestDataContainer;
  * @method string getVersionTo()
  * @method array getExpectedResult()
  */
-class LoadDataContainer extends TestDataContainer
+class CommandDataContainer extends TestDataContainer
+{
+}
+
+/**
+ * @method string getVersionFrom()
+ * @method string getVersionTo()
+ * @method array getExpectedResult()
+ */
+class LoadPackagesDataContainer extends TestDataContainer
+{
+}
+
+/**
+ * @method string getVersionFrom()
+ * @method string getVersionTo()
+ * @method array getExpectedResult()
+ */
+class LoadRuleConfigsDataContainer extends TestDataContainer
 {
 }
 
 /**
  * @method string getDirectory()
- * @method array getRules()
+ * @method array getRuleConfigs()
  * @method array getExpectedResult()
  */
 class DirectoryProcessDataContainer extends TestDataContainer
@@ -24,6 +42,7 @@ class DirectoryProcessDataContainer extends TestDataContainer
  * @method string getFilename()
  * @method array getRules()
  * @method mixed getExpectedResult()
+ * @method string getReportFilename()
  */
 class FileProcessDataContainer extends TestDataContainer
 {
@@ -45,5 +64,14 @@ class FromArrayIncorrectDataContainer extends TestDataContainer
  * @method mixed getValue()
  */
 class GettersDataContainer extends TestDataContainer
+{
+}
+
+/**
+ * @method array getInstalledVersions()
+ * @method array getRequiredVersions()
+ * @method array getExpectedResult()
+ */
+class VersionComparatorDataContainer extends TestDataContainer
 {
 }