+++ changelog.md
@@ -2,3 +2,8 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.0](https://github.com/raptor-mvk/php-migration-helper/releases/tag/v1.0.0) - 2019-09-17
+### Added
+- Command `migration-report` that checks
+  - package versions (phpunit/phpunit and symfony/monolog-bundle)
+  - potential code problems while migrating from PHP 7.1 to 7.2 and 7.3